% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_loader.R
\name{compute_phylo_distance}
\alias{compute_phylo_distance}
\title{Compute Phylogenetic Distance Matrix}
\usage{
compute_phylo_distance(phy_tree, otu_names)
}
\arguments{
\item{phy_tree}{An ape phylo object}

\item{otu_names}{Character vector of OTU names to include}
}
\value{
Numeric matrix of pairwise distances (n_otus x n_otus)
}
\description{
Computes pairwise phylogenetic distances between OTUs using the
phylogenetic tree.
}
\keyword{internal}
