% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{ensure_dir_exists}
\alias{ensure_dir_exists}
\title{Ensure Directory Exists}
\usage{
ensure_dir_exists(directory)
}
\arguments{
\item{directory}{Directory path to create}
}
\value{
Absolute path to the directory
}
\description{
Creates a directory and all parent directories if they don't exist.
}
\examples{
\dontrun{
ensure_dir_exists("outputs/models")
}

}
