% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate Model}
\usage{
evaluate_model(
  model,
  data_loader,
  device,
  k_alpha,
  k_beta,
  k_otu,
  k_time,
  k_thresh,
  k_slope,
  priors,
  is_abun_only
)
}
\arguments{
\item{model}{MDITRE model}

\item{data_loader}{torch dataloader}

\item{device}{torch device}

\item{k_alpha}{Temperature for alpha binary concrete}

\item{k_beta}{Temperature for beta binary concrete}

\item{k_otu}{Temperature for OTU selection}

\item{k_time}{Temperature for time window}

\item{k_thresh}{Temperature for threshold}

\item{k_slope}{Temperature for slope}

\item{priors}{Named list of prior parameters}

\item{is_abun_only}{Whether model is abundance-only}
}
\value{
List with metrics (total_loss, ce_loss, f1)
}
\description{
Evaluate model on a dataset (validation or test).
}
\keyword{internal}
