% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_loader.R
\name{filter_otus}
\alias{filter_otus}
\title{Filter OTUs by Prevalence and Abundance}
\usage{
filter_otus(otu_mat, min_prevalence = 0.1, min_abundance = 1e-04)
}
\arguments{
\item{otu_mat}{Matrix of OTU abundances (samples x OTUs)}

\item{min_prevalence}{Minimum fraction of samples OTU must appear in}

\item{min_abundance}{Minimum mean abundance threshold}
}
\value{
Filtered OTU matrix
}
\description{
Filter OTUs by Prevalence and Abundance
}
\keyword{internal}
