% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{get_package_root}
\alias{get_package_root}
\title{Get MDITRE Package Installation Directory}
\usage{
get_package_root()
}
\value{
Character string with absolute path to package directory,
or NULL if package is not installed
}
\description{
Returns the directory where the MDITRE R package is installed.
This could be in a user library (install.packages()) or a
development directory (devtools::load_all()).
}
\examples{
\dontrun{
pkg_dir <- get_package_root()
# Installed: "/usr/local/lib/R/site-library/rmditre"
# Dev mode: "/home/username/mditre/R"
}

}
