% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{get_platform_info}
\alias{get_platform_info}
\title{Get Platform Information}
\usage{
get_platform_info()
}
\value{
Named list with platform details
}
\description{
Returns information about the current platform.
}
\examples{
\dontrun{
info <- get_platform_info()
# Returns list with: os, os_version, r_version, sep, home, cwd
}

}
