% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{get_r_dir}
\alias{get_r_dir}
\title{Get R Implementation Directory}
\usage{
get_r_dir()
}
\value{
Character string with absolute path to R package directory
}
\description{
Returns the R package directory location. In development mode, this is
the R/ subdirectory. When installed, this is the package library location.
}
