% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{inv_transf_log}
\alias{inv_transf_log}
\title{Inverse Logarithmic Transformation}
\usage{
inv_transf_log(x, u, l)
}
\arguments{
\item{x}{Input tensor in \link{l, u}}

\item{u}{Upper bound}

\item{l}{Lower bound}
}
\value{
Transformed values (unbounded)
}
\description{
Inverse of transf_log. Transform values from \link{l, u} back to unbounded space.
}
\details{
Uses logit function (inverse sigmoid): logit(p) = log(p / (1-p))
}
\examples{
\dontrun{
library(torch)
# Forward and inverse should be identity
x <- torch_randn(10)
y <- transf_log(x, u = 1, l = 0)
x_recovered <- inv_transf_log(y, u = 1, l = 0)
}
}
