% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer1_phylogenetic_focus.R
\name{layer1_phylogenetic_focus}
\alias{layer1_phylogenetic_focus}
\title{Layer 1: Phylogenetic Focus Mechanisms}
\description{
Spatial aggregation layers that aggregate microbial time-series based on
phylogenetic distance, allowing the model to focus on taxonomically related
groups of OTUs/ASVs.
}
