% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer2_temporal_focus.R
\name{layer2_temporal_focus}
\alias{layer2_temporal_focus}
\title{Temporal Aggregation Layers for Time Focus}
\description{
These layers aggregate microbial time-series along the time dimension,
selecting contiguous time windows important for the prediction task.
}
