% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_layer.R
\name{LayerRegistry}
\alias{LayerRegistry}
\title{Layer Registry for MDITRE}
\description{
Registry for dynamically managing available layers. Allows for runtime
layer selection, easy addition of new layer implementations, and version
control of layer implementations.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{registry}}{Named list storing registered layer classes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LayerRegistry-register}{\code{LayerRegistry$register()}}
\item \href{#method-LayerRegistry-get}{\code{LayerRegistry$get()}}
\item \href{#method-LayerRegistry-list_layers}{\code{LayerRegistry$list_layers()}}
\item \href{#method-LayerRegistry-get_info}{\code{LayerRegistry$get_info()}}
\item \href{#method-LayerRegistry-remove}{\code{LayerRegistry$remove()}}
\item \href{#method-LayerRegistry-clone}{\code{LayerRegistry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LayerRegistry-register"></a>}}
\if{latex}{\out{\hypertarget{method-LayerRegistry-register}{}}}
\subsection{Method \code{register()}}{
Register a new layer type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayerRegistry$register(name, layer_class, version = "1.0.0")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string name for the layer}

\item{\code{layer_class}}{The nn_module class definition}

\item{\code{version}}{Character string version (default: "1.0.0")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LayerRegistry-get"></a>}}
\if{latex}{\out{\hypertarget{method-LayerRegistry-get}{}}}
\subsection{Method \code{get()}}{
Get a registered layer class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayerRegistry$get(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string name of the layer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The layer class definition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LayerRegistry-list_layers"></a>}}
\if{latex}{\out{\hypertarget{method-LayerRegistry-list_layers}{}}}
\subsection{Method \code{list_layers()}}{
List all registered layers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayerRegistry$list_layers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character vector of registered layer names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LayerRegistry-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-LayerRegistry-get_info}{}}}
\subsection{Method \code{get_info()}}{
Get information about a registered layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayerRegistry$get_info(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string name of the layer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named list with layer metadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LayerRegistry-remove"></a>}}
\if{latex}{\out{\hypertarget{method-LayerRegistry-remove}{}}}
\subsection{Method \code{remove()}}{
Remove a layer from registry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayerRegistry$remove(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string name of the layer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LayerRegistry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LayerRegistry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LayerRegistry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
