% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{negbin_loss}
\alias{negbin_loss}
\title{Negative Binomial Loss}
\usage{
negbin_loss(x, mean, variance)
}
\arguments{
\item{x}{Observed count}

\item{mean}{Mean parameter}

\item{variance}{Variance parameter}
}
\value{
Negative log-likelihood
}
\description{
Compute negative log-likelihood of negative binomial distribution.
Used as a prior to encourage sparsity in the number of active components.
}
\keyword{internal}
