% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_loader.R
\name{normalize_abundance}
\alias{normalize_abundance}
\title{Normalize OTU Abundances}
\usage{
normalize_abundance(otu_mat)
}
\arguments{
\item{otu_mat}{Matrix of OTU abundances (samples x OTUs)}
}
\value{
Normalized OTU matrix
}
\description{
Performs total sum scaling (relative abundance) normalization.
}
\keyword{internal}
