% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_loader.R
\name{organize_by_subject}
\alias{organize_by_subject}
\title{Organize Data by Subject}
\usage{
organize_by_subject(otu_mat, subjects, times, labels)
}
\arguments{
\item{otu_mat}{Matrix of OTU abundances (samples x OTUs)}

\item{subjects}{Vector of subject IDs for each sample}

\item{times}{Vector of time points for each sample}

\item{labels}{Vector of labels for each sample}
}
\value{
List with X (3D array), y (vector), times (matrix), mask (matrix)
}
\description{
Reshapes flat sample x OTU matrix into subject x OTU x time tensor.
}
\keyword{internal}
