% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot Confusion Matrix}
\usage{
plot_confusion_matrix(metrics, save_path = NULL, width = 6, height = 6)
}
\arguments{
\item{metrics}{Metrics object from compute_metrics()}

\item{save_path}{Optional path to save plot}

\item{width}{Plot width in inches (default: 6)}

\item{height}{Plot height in inches (default: 6)}
}
\value{
ggplot object (invisibly)
}
\description{
Visualize confusion matrix as a heatmap.
}
\examples{
\dontrun{
metrics <- compute_metrics(predictions, labels)
plot_confusion_matrix(metrics)
}

}
