% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Plot Model Comparison}
\usage{
plot_model_comparison(
  comparison_results,
  metric = "f1",
  save_path = NULL,
  width = 10,
  height = 6
)
}
\arguments{
\item{comparison_results}{Results from compare_models()}

\item{metric}{Metric to compare (default: "f1")}

\item{save_path}{Optional path to save plot}

\item{width}{Plot width in inches (default: 10)}

\item{height}{Plot height in inches (default: 6)}
}
\value{
ggplot object (invisibly)
}
\description{
Compare performance of multiple models.
}
\examples{
\dontrun{
comparison <- compare_models(mditre_data, model_configs)
plot_model_comparison(comparison)
}

}
