% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_parameter_distributions}
\alias{plot_parameter_distributions}
\title{Plot Parameter Distributions}
\usage{
plot_parameter_distributions(
  model,
  parameters = c("kappa", "eta", "thresh"),
  save_path = NULL,
  width = 10,
  height = 6
)
}
\arguments{
\item{model}{Trained MDITRE model}

\item{parameters}{Vector of parameter names to plot
(e.g., c("kappa", "eta", "thresh"))}

\item{save_path}{Optional path to save plot}

\item{width}{Plot width in inches (default: 10)}

\item{height}{Plot height in inches (default: 6)}
}
\value{
ggplot object (invisibly)
}
\description{
Visualize distributions of learned model parameters.
}
\examples{
\dontrun{
# After training
plot_parameter_distributions(trained_model, c("kappa", "eta", "thresh"))
}

}
