% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_roc_curve}
\alias{plot_roc_curve}
\title{Plot ROC Curve}
\usage{
plot_roc_curve(
  predictions,
  labels,
  title = "ROC Curve",
  save_path = NULL,
  width = 6,
  height = 6
)
}
\arguments{
\item{predictions}{Predicted probabilities}

\item{labels}{True binary labels}

\item{title}{Plot title (default: "ROC Curve")}

\item{save_path}{Optional path to save plot}

\item{width}{Plot width in inches (default: 6)}

\item{height}{Plot height in inches (default: 6)}
}
\value{
ggplot object (invisibly)
}
\description{
Plot ROC curve with AUC.
}
\examples{
\dontrun{
# After evaluation
eval_result <- evaluate_model_on_data(model, test_loader, return_predictions = TRUE)
plot_roc_curve(eval_result$predictions, eval_result$labels)
}

}
