% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeding.R
\name{print.mditre_seed_generator}
\alias{print.mditre_seed_generator}
\title{Print Method for MDITRE Seed Generator}
\usage{
\method{print}{mditre_seed_generator}(x, ...)
}
\arguments{
\item{x}{An mditre_seed_generator object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for MDITRE Seed Generator
}
