% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_loader.R
\name{print_mditre_data_summary}
\alias{print_mditre_data_summary}
\title{Print Summary of MDITRE Data}
\usage{
print_mditre_data_summary(mditre_data)
}
\arguments{
\item{mditre_data}{List returned from \code{phyloseq_to_mditre}}
}
\description{
Print Summary of MDITRE Data
}
