% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{print_metrics}
\alias{print_metrics}
\title{Print Metrics Summary}
\usage{
print_metrics(metrics)
}
\arguments{
\item{metrics}{Metrics object from compute_metrics()}
}
\description{
Pretty print classification metrics.
}
