% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{save_checkpoint}
\alias{save_checkpoint}
\title{Save Model Checkpoint}
\usage{
save_checkpoint(model, checkpoint_path)
}
\arguments{
\item{model}{MDITRE model}

\item{checkpoint_path}{Path to save checkpoint}
}
\description{
Save model state to a checkpoint file.
}
\examples{
\dontrun{
save_checkpoint(model, "checkpoints/my_model.pt")
}

}
