% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_utils.R
\name{transf_log}
\alias{transf_log}
\title{Logarithmic Transformation with Bounds}
\usage{
transf_log(x, u, l)
}
\arguments{
\item{x}{Input tensor}

\item{u}{Upper bound}

\item{l}{Lower bound}
}
\value{
Transformed values in \link{l, u}
}
\description{
Transform values from unbounded space to bounded interval \link{l, u}
using sigmoid function.
}
\examples{
\dontrun{
library(torch)
x <- torch_randn(10)
# Transform to [0, 1]
y <- transf_log(x, u = 1, l = 0)
}
}
